comment added
tried to do the search by file id, and taking posts from feed by the id. but deleting from search doesn't delete from feed, until refreshed.

+++ deletepost.php
@@ -12,12 +12,19 @@ if(isset($_POST['id']) && !empty($_POST['id'])){
     $row=mysqli_fetch_row($query_run);
     $file=$row[0];
     $groupid=$row[1];
-    $yes=@unlink('C:/xampp/htdocs/GroupStudy/GroupProject/files/'.$groupid.'_'.$file);
-    if($yes==false) die('Server Error. Post could not be deleted');
+    
+    if($file!=""){
+        $yes=@unlink('C:/xampp/htdocs/GroupStudy/GroupProject/files/'.$groupid.'_'.$file);
+        if($yes==false) die('Server Error. Post could not be deleted');
+    }
+    
     $query="delete from files where id='".$id."'";
     if(!mysqli_query($con,$query)) die('Server Error');
     
     $query="delete from tags where file_id='".$id."'";
+    if(!mysqli_query($con,$query)) die('Server Error');
+
+    $query="delete from comments where file_id='".$id."'";
     if(!mysqli_query($con,$query)) die('Server Error');
 }
 else die('Error');
